refactor(mocks): restructure mocks to avoid calling msw's setupWorker in Node
